Added key functionality
I also added two important variables: turnPhase and currentTurn.
currentTurn is pretty self-explanatory, and was added because a similar variable exists in real Jevil fight. I haven't done anything with it yet, but it will be useful once we start working on Jevil's attacks.
turnPhase keeps track of which part of the turn is currently active, cycling every turn. These numbers might not be final, but what I have planned for now is as such: 0-2 is for selecting each character's actions, 3-5 has the characters performing those actions (skipping downed or attacking characters as needed), 6 is for the characters who are attacking, 7 is Jevil's text line, and 8 is the bullet hell section.